+++ src/dg/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(DG_SOURCE
     high_order_grid.cpp
+    meshmover_linear_elasticity.cpp
     dg.cpp
     drdx.cpp
     residual_sparsity_patterns.cpp
