+++ src/dg/high_order_grid.h
@@ -293,70 +293,6 @@ public:
     virtual dealii::UpdateFlags get_needed_update_flags () const override;
 };
 
-namespace MeshMover
-{
-    template <int dim = PHILIP_DIM, typename real = double, typename VectorType = dealii::LinearAlgebra::distributed::Vector<double>, typename DoFHandlerType = dealii::DoFHandler<PHILIP_DIM>>
-    class LinearElasticity
-    {
-#if PHILIP_DIM==1 // dealii::parallel::distributed::Triangulation<dim> does not work for 1D
-    /** Triangulation to store the grid.
-     *  In 1D, dealii::Triangulation<dim> is used.
-     *  In 2D, 3D, dealii::parallel::distributed::Triangulation<dim> is used.
-     */
-    using Triangulation = dealii::Triangulation<dim>;
-#else
-    /** Triangulation to store the grid.
-     *  In 1D, dealii::Triangulation<dim> is used.
-     *  In 2D, 3D, dealii::parallel::distributed::Triangulation<dim> is used.
-     */
-    using Triangulation = dealii::parallel::distributed::Triangulation<dim>;
-#endif
-      public:
-        LinearElasticity(
-            const HighOrderGrid<dim,real,VectorType,DoFHandlerType> &high_order_grid,
-			const dealii::LinearAlgebra::distributed::Vector<double> &boundary_displacements_vector);
-        ~LinearElasticity();
-        VectorType get_volume_displacements();
-
-		void evaluate_dXvdXs();
-
-        VectorType displacement_solution;
-        dealii::AffineConstraints<double> hanging_node_constraints;
-        //std::vector<dealii::TrilinosWrappers::MPI::Vector> dXvdXs;
-        std::vector<dealii::LinearAlgebra::distributed::Vector<double>> dXvdXs;
-      private:
-        void setup_system();
-        void assemble_system();
-        void solve_timestep();
-        unsigned int solve_linear_problem();
-        void move_mesh();
-        void setup_quadrature_point_history();
-        const Triangulation &triangulation;
-        dealii::FESystem<dim> fe;
-        std::shared_ptr<dealii::MappingFEField<dim,dim,VectorType,DoFHandlerType>> mapping_fe_field;
-        DoFHandlerType dof_handler;
-        const dealii::QGauss<dim> quadrature_formula;
-        dealii::TrilinosWrappers::SparseMatrix system_matrix;
-        dealii::TrilinosWrappers::MPI::Vector system_rhs;
-        dealii::TrilinosWrappers::SparseMatrix system_matrix_unconstrained;
-        dealii::TrilinosWrappers::MPI::Vector system_rhs_unconstrained;
-
-        dealii::AffineConstraints<double> all_constraints;
-        dealii::AffineConstraints<double> dirichlet_boundary_constraints;
-
-        MPI_Comm mpi_communicator;
-        const unsigned int n_mpi_processes;
-        const unsigned int this_mpi_process;
-        dealii::ConditionalOStream pcout;
-        std::vector<dealii::types::global_dof_index> local_dofs_per_process;
-        dealii::IndexSet locally_owned_dofs;
-        dealii::IndexSet locally_relevant_dofs;
-
-        const dealii::LinearAlgebra::distributed::Vector<int> &boundary_ids_vector;
-        const dealii::LinearAlgebra::distributed::Vector<double> &boundary_displacements_vector;
-
-    };
-} // namespace MeshMover
 
 } // namespace PHiLiP
