+++ tests/unit_tests/grid/LinearElasticity_mesh_movement.cpp
@@ -24,6 +24,7 @@
 #include <deal.II/fe/mapping_q.h> 
 
 #include "dg/high_order_grid.h"
+#include "dg/meshmover_linear_elasticity.hpp"
 #include "parameters/all_parameters.h"
 
 template<int dim>
